+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.10.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.10.1...@rudderstack/analytics-js-plugins@3.10.2) (2025-06-20)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.21.0`
+* `@rudderstack/analytics-js-cookies` updated to version `0.4.27`
 ## [3.10.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.10.0...@rudderstack/analytics-js-plugins@3.10.1) (2025-06-12)