+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.51.4](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.51.3...rudder-sdk-js@2.51.4) (2025-06-20)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.21.0`
 ## [2.51.3](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.51.2...rudder-sdk-js@2.51.3) (2025-06-11)
 
 ### Dependency Updates